requirements.txt
torch
transformers
diffusers
peft
accelerate
Pillow
PyYAML
matplotlib
numpy
torchmetrics
open-clip-torch